move OsStr helpers to a separate file
